Move class to a separate file
