+++ src/client/datascience/jupyter/kernels/kernelSelector.ts
@@ -23,7 +23,8 @@ import { sendNotebookOrKernelLanguageTelemetry } from '../../common';
 import { Commands, Settings, Telemetry } from '../../constants';
 import { IKernelFinder } from '../../kernel-launcher/types';
 import { isPythonNotebook } from '../../notebook/helpers/helpers';
-import { getInterpreterInfoStoredInMetadata, PreferredRemoteKernelIdProvider } from '../../notebookStorage/baseModel';
+import { getInterpreterInfoStoredInMetadata } from '../../notebookStorage/baseModel';
+import { PreferredRemoteKernelIdProvider } from '../../notebookStorage/preferredRemoteKernelIdProvider';
 import { reportAction } from '../../progress/decorator';
 import { ReportableAction } from '../../progress/types';
 import {
@@ -76,7 +77,8 @@ export class KernelSelector implements IKernelSelectionUsage {
         @inject(IConfigurationService) private configService: IConfigurationService,
         @inject(IDisposableRegistry) disposableRegistry: IDisposableRegistry,
         @inject(IPythonExtensionChecker) private readonly extensionChecker: IPythonExtensionChecker,
-        @inject(PreferredRemoteKernelIdProvider) private readonly preferredRemoteKernelIdProvider: PreferredRemoteKernelIdProvider
+        @inject(PreferredRemoteKernelIdProvider)
+        private readonly preferredRemoteKernelIdProvider: PreferredRemoteKernelIdProvider
     ) {
         disposableRegistry.push(
             this.jupyterSessionManagerFactory.onRestartSessionCreated(this.addKernelToIgnoreList.bind(this))
@@ -260,7 +262,9 @@ export class KernelSelector implements IKernelSelectionUsage {
         ]);
 
         // First check for a live active session.
-        const preferredKernelId = resource ? this.preferredRemoteKernelIdProvider.getPreferredRemoteKernelId(resource) : undefined;
+        const preferredKernelId = resource
+            ? this.preferredRemoteKernelIdProvider.getPreferredRemoteKernelId(resource)
+            : undefined;
         if (preferredKernelId) {
             const session = sessions?.find((s) => s.kernel.id === preferredKernelId);
             if (session) {
